+++ CHANGELOG.md
@@ -47,6 +47,7 @@ Compatibility:
 * `File#path` now returns a new mutable String on every call like MRI (#2115).
 * Avoid infinite recursion when redefining `Warning#warn` and calling `Kernel#warn` (#2109).
 * Convert objects with `#to_path` in `$LOAD_PATH` (#2119).
+* Handle the functions being native for `rb_thread_call_without_gvl()` (#2090).
 
 Performance: